finished test refactor
